+++ pyodide-build/pyodide_build/pywasmcross.py
@@ -329,6 +329,9 @@ def handle_command(line, args, dryrun=False):
             continue
         if arg == "-lffi":
             continue
+        # This flag is needed to build numpy with SIMD optimization
+        if arg == "-mpopcnt":
+            continue
         # The native build is possibly multithreaded, but the emscripten one
         # definitely isn't
         arg = re.sub(r"/python([0-9]\.[0-9]+)m", r"/python\1", arg)